requirements.txt
# ── Core Framework ──────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic[email]==2.7.1
pydantic-settings==2.2.1

# ── Database (async SQLAlchemy + PostgreSQL) ─────────────────────────────────
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9          # used by Alembic sync env.py

# ── Auth ─────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9         # form parsing for login

# ── AWS (S3 signed URLs, SQS, SNS) ──────────────────────────────────────────
boto3==1.34.93
aiobotocore==2.13.0             # async boto3 wrapper

# ── Background / messaging ───────────────────────────────────────────────────
celery==5.4.0
redis==5.0.4
kombu==5.3.5

# ── Testing ──────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0                   # async test client for FastAPI
factory-boy==3.3.0
Faker==24.11.0
pytest-cov==5.0.0

# ── Utilities ────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
structlog==24.1.0               # structured JSON logging
tenacity==8.3.0                 # retry logic
pillow==10.3.0                  # image validation on upload

# ── Rate limiting ─────────────────────────────────────────────────────────────
slowapi==0.1.9                  # Starlette/FastAPI rate limiter (slowapi wraps limits)
limits==3.10.1                  # dependency of slowapi

# ── Email (SMTP async) ────────────────────────────────────────────────────────
aiosmtplib==3.0.1               # async SMTP client (SES SMTP endpoint)

# ── Lambda ASGI adapter ───────────────────────────────────────────────────────
mangum==0.17.0                  # AWS Lambda + API Gateway → ASGI bridge
